toolkit/components/passwordmgr/test/test_zzz_finish.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test finalization for Login Manager</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: finalization.
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: finalization **/


// Get the pwmgr service
var Cc_pwmgr = SpecialPowers.Cc["@mozilla.org/login-manager;1"];
ok(Cc_pwmgr != null, "Access Cc[@mozilla.org/login-manager;1]");

var Ci_pwmgr = SpecialPowers.Ci.nsILoginManager;
ok(Ci_pwmgr != null, "Access Ci.nsILoginManager");

var pwmgr = Cc_pwmgr.getService(Ci_pwmgr);
ok(pwmgr != null, "pwmgr getService()");


// Remove all logins, so future test runs start off clean.
pwmgr.removeAllLogins();

var logins = pwmgr.getAllLogins();
ok(logins != null, "getAllLogins()");
is(logins.length, 0, "ensure no remaining logins");

// Remove any disabled hosts
var disabledHosts = pwmgr.getAllDisabledHosts();
ok(disabledHosts != null, "getAllDisabledHosts()");
disabledHosts.forEach(host => pwmgr.setLoginSavingEnabled(host, true));

disabledHosts = pwmgr.getAllDisabledHosts();
ok(disabledHosts != null, "getAllDisabledHosts()");
is(disabledHosts.length, 0, "ensure no remaining disabled hosts");

</script>
</pre>
</body>
</html>
